requirements.txt
discord
colorama
python-dotenv
requests